requirements.txt
torch>=2.9.0
torchvision>=0.18.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
scikit-image>=0.22.0
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
nltk>=3.8.1
transformers>=4.30.0
opencv-python>=4.8.0
scipy>=1.11.0
pandas>=2.0.0
